adding separate files for potential,        species and mechanics solves
