--- Source/Kernels_3d.H
@@ -1,11 +1,0 @@
-#ifndef Kernels_H_
-#define Kernels_H_
-
-#include <Prob.H>
-#include <advance_3d.H>
-#include <compute_flux_3d.H>
-#include <echemAMR_Tagging.H>
-#include <echemAMR_BCFill.H>
-#include <bv_utils.H>
-
-#endif
