+++ Source/echemAMR_Integrate.cpp
@@ -1,6 +1,5 @@
 #include <AMReX_MultiFabUtil.H>
 #include <AMReX_FillPatchUtil.H>
-
 #include <echemAMR.H>
 #include <Chemistry.H>
 #include <IntegralUtils.H>
